Building Machine Learning Models to Predict Credit Card Fraud
This machine learning project is focused on building an effective model to help detect fraudulent credit card transactions. The goal of this project is to explore the given dataset, perform the necessary visualisations to understand the dataset, scale the features provided in the dataset, build machine learning models for classifying transactions, and comparing the models.